+++ lib-rdm/src/rdmdeviceprint.cpp
@@ -29,24 +29,10 @@
 
 void RDMDevice::Print(void) {
 	printf("RDM Device configuration\n");
-	printf(" Manufacturer Name : %.*s\n",
-			static_cast<int>(m_tRDMDevice.nDdeviceManufacturerNameLength),
-			m_tRDMDevice.aDeviceManufacturerName);
-	printf(" Manufacturer ID   : %.2X%.2X\n",
-			static_cast<int>(m_tRDMDevice.aDeviceUID[0]),
-			static_cast<int>(m_tRDMDevice.aDeviceUID[1]));
-	printf(" Serial Number     : %.2X%.2X%.2X%.2X\n",
-			static_cast<int>(m_tRDMDevice.aDeviceSN[3]),
-			static_cast<int>(m_tRDMDevice.aDeviceSN[2]),
-			static_cast<int>(m_tRDMDevice.aDeviceSN[1]),
-			static_cast<int>(m_tRDMDevice.aDeviceSN[0]));
-	printf(" Root label        : %.*s\n",
-			static_cast<int>(m_tRDMDevice.nDeviceRootLabelLength),
-			m_tRDMDevice.aDeviceRootLabel);
-	printf(" Product Category  : %.2X%.2X\n",
-			static_cast<int>(m_tRDMDevice.nProductCategory) >> 8,
-			static_cast<int>(m_tRDMDevice.nProductCategory) & 0xFF);
-	printf(" Product Detail    : %.2X%.2X\n",
-			static_cast<int>(m_tRDMDevice.nProductDetail) >> 8,
-			static_cast<int>(m_tRDMDevice.nProductDetail) & 0xFF);
+	printf(" Manufacturer Name : %.*s\n", m_tRDMDevice.nDdeviceManufacturerNameLength, m_tRDMDevice.aDeviceManufacturerName);
+	printf(" Manufacturer ID   : %.2X%.2X\n", m_tRDMDevice.aDeviceUID[0], m_tRDMDevice.aDeviceUID[1]);
+	printf(" Serial Number     : %.2X%.2X%.2X%.2X\n", m_tRDMDevice.aDeviceSN[3], m_tRDMDevice.aDeviceSN[2], m_tRDMDevice.aDeviceSN[1], m_tRDMDevice.aDeviceSN[0]);
+	printf(" Root label        : %.*s\n", m_tRDMDevice.nDeviceRootLabelLength, m_tRDMDevice.aDeviceRootLabel);
+	printf(" Product Category  : %.2X%.2X\n", m_tRDMDevice.nProductCategory >> 8, m_tRDMDevice.nProductCategory & 0xFF);
+	printf(" Product Detail    : %.2X%.2X\n", m_tRDMDevice.nProductDetail >> 8, m_tRDMDevice.nProductDetail & 0xFF);
 }
