+++ lib-showfile/src/showfiletftp.cpp
@@ -25,7 +25,6 @@
 
 #include <stdint.h>
 #include <stdio.h>
-#include <assert.h>
 
 #include "showfiletftp.h"
 #include "showfile.h"
@@ -53,7 +52,7 @@ void ShowFileTFTP::Exit(void) {
 }
 
 bool ShowFileTFTP::FileOpen(const char *pFileName, TTFTPMode tMode) {
-	DEBUG_PRINTF("pFileName=%s, tMode=%d", pFileName, static_cast<int>(tMode));
+	DEBUG_PRINTF("pFileName=%s, tMode=%d", pFileName, tMode);
 
 	uint8_t nShowFileNumber;
 	if (!ShowFile::CheckShowFileName(pFileName, nShowFileNumber)) {
@@ -66,7 +65,7 @@ bool ShowFileTFTP::FileOpen(const char *pFileName, TTFTPMode tMode) {
 }
 
 bool ShowFileTFTP::FileCreate(const char *pFileName, TTFTPMode tMode) {
-	DEBUG_PRINTF("pFileName=%s, tMode=%d", pFileName, static_cast<int>(tMode));
+	DEBUG_PRINTF("pFileName=%s, tMode=%d", pFileName, tMode);
 
 	uint8_t nShowFileNumber;
 	if (!ShowFile::CheckShowFileName(pFileName, nShowFileNumber)) {
