+++ lib-spiflashstore/src/storewidget.cpp
@@ -71,7 +71,7 @@ void StoreWidget::Copy(struct TWidgetParams* pWidgetParams) {
 void StoreWidget::UpdateBreakTime(uint8_t nBreakTime) {
 	DEBUG_ENTRY
 
-	DEBUG_PRINTF("offsetof=%d", (int) __builtin_offsetof(struct TWidgetParams, nBreakTime));
+	DEBUG_PRINTF("offsetof=%d", __builtin_offsetof(struct TWidgetParams, nBreakTime));
 
 	SpiFlashStore::Get()->Update(STORE_WIDGET, __builtin_offsetof(struct TWidgetParams, nBreakTime), &nBreakTime, sizeof(uint8_t), WIDGET_PARAMS_MASK_BREAK_TIME);
 
@@ -81,7 +81,7 @@ void StoreWidget::UpdateBreakTime(uint8_t nBreakTime) {
 void StoreWidget::UpdateMabTime(uint8_t nMabTime) {
 	DEBUG_ENTRY
 
-	DEBUG_PRINTF("offsetof=%d", (int) __builtin_offsetof(struct TWidgetParams, nMabTime));
+	DEBUG_PRINTF("offsetof=%d", __builtin_offsetof(struct TWidgetParams, nMabTime));
 
 	SpiFlashStore::Get()->Update(STORE_WIDGET, __builtin_offsetof(struct TWidgetParams, nMabTime), &nMabTime, sizeof(uint8_t), WIDGET_PARAMS_MASK_MAB_TIME);
 
@@ -91,7 +91,7 @@ void StoreWidget::UpdateMabTime(uint8_t nMabTime) {
 void StoreWidget::UpdateRefreshRate(uint8_t nRefreshRate) {
 	DEBUG_ENTRY
 
-	DEBUG_PRINTF("offsetof=%d", (int) __builtin_offsetof(struct TWidgetParams, nRefreshRate));
+	DEBUG_PRINTF("offsetof=%d", __builtin_offsetof(struct TWidgetParams, nRefreshRate));
 
 	SpiFlashStore::Get()->Update(STORE_WIDGET, __builtin_offsetof(struct TWidgetParams, nRefreshRate), &nRefreshRate, sizeof(uint8_t), WIDGET_PARAMS_MASK_REFRESH_RATE);
